pyproject.toml
[project]
name = "azure-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "azure-ai-evaluation>=1.13.7",
    "azure-ai-ml>=1.30.0",
    "azure-core>=1.37.0",
    "azure-identity>=1.25.1",
    "azure-storage-blob>=12.19.0",
    "azureml-mlflow>=1.61.0.post1",
    "cryptography>=42.0.2",
    "lagom>=2.7.7",
    "mlflow>=3.2.1",
    "openai>=2.11.0",
    "pre-commit>=4.5.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "urllib3>=2.6.2",
]

[dependency-groups]
dev = [
    "pip-audit>=2.10.0",
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.9",
]

[tool.ruff]
exclude = ["typings", ".venv"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]

[tool.coverage.report]
fail_under = 90

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "azure_python/protocols/*.py",
    "azure_python/hosting.py"
]
